Rename `schnorr` module to `bip340`
The name better describes the exact signature scheme.
